2008-09-11 18:00 UTC-0800 Pritpal Bedi (pritpal@vouchcac.com)   * contrib/gtwvg/wvgutils.c     ! HB_FUNC( WVT__MAKEDLGTEMPLATE )          hb_retclen( ( LPSTR ) pdlgtemplate, ( ULONG ) ( p - pdlgtemplate ) );    =>          hb_retclen( ( LPSTR ) pdlgtemplate, ( ( ULONG ) p - ( ULONG ) pdlgtemplate ) ) ;       Fix to open modeless/modal dialogs.       ; NOTE: Need to be fixed in v1.0.0 also.

+++ harbour/contrib/gtwvg/wvgutils.c
@@ -1288,7 +1288,7 @@ HB_FUNC( WVT__MAKEDLGTEMPLATE )
 
     p = lpwAlign( p );
 
-    hb_retclen( ( LPSTR ) pdlgtemplate, ( ULONG ) ( p - pdlgtemplate ) );
+    hb_retclen( ( LPSTR ) pdlgtemplate, ( ( ULONG ) p - ( ULONG ) pdlgtemplate ) ) ;
 
     LocalFree( LocalHandle( pdlgtemplate ) );
 }
@@ -1301,6 +1301,7 @@ HB_FUNC( WVT__MAKEDLGTEMPLATE )
 HB_EXPORT LPWORD lpwAlign( LPWORD lpIn )
 {
    HB_PTRDIFF ul = ( HB_PTRDIFF ) lpIn;
+
    ul += 3;
    ul >>=2;
    ul <<=2;
